Move Markdown rendering engine over to Remark, and some misc visual tweaks
- Add clock icon next to reading time
- Use sans serif font for blog post date and bylines
- Fix <code> tags in links via Rehype rather than with JS
- Add .cjs suffix to Eleventy config file
