+++ package.json
@@ -5,8 +5,8 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "clean": "rm -rf dist",
-    "eleventy:default": "eleventy",
-    "eleventy:watch": "eleventy --watch",
+    "eleventy:default": "eleventy --config=.eleventy.cjs",
+    "eleventy:watch": "eleventy --watch --config=.eleventy.cjs",
     "webpack:prod": "webpack --mode=production",
     "webpack:dev": "webpack --mode=development",
     "serve": "light-server -c .lightserverrc",
@@ -33,23 +33,24 @@
     "release-it": "^13.6.2"
   },
   "dependencies": {
-    "@11ty/eleventy": "^0.11.1",
+    "@11ty/eleventy": "^0.12.1",
     "@11ty/eleventy-navigation": "^0.1.5",
     "@11ty/eleventy-plugin-rss": "^1.0.7",
+    "@amanda-mitchell/remark-renumber-footnotes": "^1.0.0",
     "@ap-spectrum-web-components/slider": "file:local_modules/slider/",
     "@babel/core": "^7.6.4",
     "@babel/plugin-proposal-class-properties": "^7.10.1",
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.6.3",
-    "@iktakahiro/markdown-it-katex": "^3.1.0",
+    "@fec/eleventy-plugin-remark": "^2.2.1",
     "@observablehq/runtime": "^4.7.0",
+    "@silvenon/remark-smartypants": "^1.0.0",
     "@spectrum-web-components/theme": "^0.4.0",
     "ap-spectrum-web-components-slider": "file:local_modules/slider/",
     "autoprefixer": "^9.6.5",
     "babel-loader": "^8.0.6",
     "breakpoint-sass": "^2.7.1",
     "commonmark-extract-text": "^0.1.0",
-    "css-has-pseudo": "^1.0.0",
     "css-loader": "^3.5.3",
     "cssnano-webpack-plugin": "^1.0.3",
     "dotenv": "^8.2.0",
@@ -59,6 +60,8 @@
     "github-emoji": "^1.1.1",
     "github-textmate-theme": "git+https://github.com/primer/github-textmate-theme.git",
     "global": "^4.4.0",
+    "hast-util-select": "4.0.2",
+    "hastscript": "6.0.0",
     "html-minifier": "^4.0.0",
     "imagemin-pngquant": "^9.0.0",
     "import-http": "^0.3.1",
@@ -66,14 +69,6 @@
     "katex": "^0.11.1",
     "linkinator": "^2.1.1",
     "luxon": "^1.24.1",
-    "markdown-it": "^11.0.0",
-    "markdown-it-admonition": "^1.0.4",
-    "markdown-it-attrs": "^3.0.2",
-    "markdown-it-div": "^1.1.0",
-    "markdown-it-emoji": "^1.4.0",
-    "markdown-it-figure": "^0.2.0",
-    "markdown-it-footnote": "^3.0.2",
-    "markdown-it-shiki": "^0.1.1",
     "mini-css-extract-plugin": "^0.9.0",
     "normalize.css": "^8.0.1",
     "npm-run-all": "^4.1.5",
@@ -81,6 +76,14 @@
     "postcss-loader": "^3.0.0",
     "puppeteer": "2.1.1",
     "read-time-estimate": "^0.0.2",
+    "rehype-raw": "^5.1.0",
+    "rehype-stringify": "^8.0.0",
+    "remark-code-frontmatter": "^1.0.0",
+    "remark-directive": "^1.0.1",
+    "remark-emoji": "^2.2.0",
+    "remark-footnotes": "^3.0.0",
+    "remark-gfm": "^1.0.0",
+    "remark-rehype": "^8.1.0",
     "resolve-url-loader": "^3.1.1",
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
@@ -91,8 +94,9 @@
     "svg-sprite": "^1.5.0",
     "synchronized-promise": "^0.3.1",
     "turbolinks": "^5.2.0",
-    "unist-builder": "^2.0.3",
-    "unist-util-visit": "^2.0.2",
+    "unist-builder": "2.0.3",
+    "unist-util-find": "^1.0.2",
+    "unist-util-visit": "2.0.3",
     "vinyl": "^2.2.0",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.9"
