+++ src/assets/styles/components/_post_list.scss
@@ -4,7 +4,7 @@
   &__year {
     color: $oc-gray-6;
     font-size: 1.5rem;
-    font-weight: normal;
+    font-weight: 700;
   }
 
   &__post {
@@ -17,9 +17,10 @@
   }
 
   &__published_at {
+    font-family: $font-family-sans-serif;
     width: 160px;
     flex-shrink: 0;
-    font-size: 1.1rem;
+    font-size: 1rem;
     color: $oc-gray-6;
     margin-right: 2.5rem;
     padding-top: 2px;
