+++ src/assets/styles/components/_talk_list.scss
@@ -38,6 +38,7 @@
   }
 
   &__outlet {
+    font-family: $font-family-sans-serif;
     font-weight: normal;
     font-size: 1.1rem;
     color: $text-color-light;
