+++ src/blog/2020-06-07-psa-dont-use-cssnano-with-postcss.md
@@ -8,9 +8,10 @@ While redesigning this site, I ran into an odd problem: my build pipeline was ge
 
 I employed the old developer standby of turning off all the possible causes one by one, and tracked the problem down to one line in my [PostCSS](https://postcss.org/) config:
 
-::: data-filename=postcss.config.js
-
 ```js
+---
+filename: postcss.config.js
+---
 module.exports = {
   plugins: {
     'postcss-import': {},
@@ -20,17 +21,16 @@ module.exports = {
 }
 ```
 
-:::
-
 At the time of writing the documentation for PostCSS[^1] and cssnano[^2] both seem to suggest the two should play nicely together. However, it turns out that according [this Github issue](https://github.com/cssnano/cssnano/issues/659) the developers of cssnano actually don't recommend loading cssnano with `postcss-loader` anymore:
 
 > Don't use cssnano in postcss-loader [...] using on postcss-loader makes impossible same optimization, also css-loader can rewrite some css stuff
 
 The issue mentions a plugin called `optimize-cssnano-plugin` but that seems to be unmaintained - a much better alternative seems to be [`cssnano-webpack-plugin`](https://www.npmjs.com/package/cssnano-webpack-plugin):
 
-::: data-filename=webpack.config.js
-
 ```js
+---
+filename: webpack.config.js
+---
 const CssnanoPlugin = require('cssnano-webpack-plugin')
 
 module.exports = {
@@ -49,8 +49,6 @@ module.exports = {
   ]
 }
 ```
-
-:::
 
 _Voilà_ - CSS compression without mangled source maps.
